// src/models/template.ts
export enum Command {
  Table,
  Dialog,
  Rules,
  Form,
  Columns
}

export enum SPACE {
  S1 = "  ",
  S2 = "    ",
  S3 = "      ",
  S4 = "        ",
  S5 = "          ",
  S6 = "            ",
  S7 = "              ",
}

export enum Component {
  Input = "input",
  InputNumbser = "inputNumbser",
  DatePicker = "datePicker",
  Switch = "switch",
}

export enum JavaType {
  String = "String", 
  Object = "Object",
  Integer = "Integer", 
  int = "int", 
  Double = "Double", 
  double = "double", 
  BigDecimal = "BigDecimal",
  LocalDate = "LocalDate", 
  OffsetDateTime = "OffsetDateTime",
  Boolean = "Boolean", 
  boolean = "boolean"
}